Move CtQ graph generation to separate file
